Add function to render time to the DOM
We refactored the DOM update method as it was producing an element every
time it ran which was A LOT, we now have it appending the time to
a single element. We created a requestAnimationFrame to animate the
changing time that also allows us to clear the interval.

We also refactored out tests to check not the time but right value is
being produced